+++ nuget-client/BrevgeneratorKlient.Tests/BrevgeneratorKlientTests.cs
@@ -1,8 +1,8 @@
 using System.Net;
-using Arbeidstilsynet.Brevgenerator.Klient.Model;
+using Arbeidstilsynet.Brevgenerator.Client.Model;
 using Xunit;
 
-namespace Arbeidstilsynet.Brevgenerator.Klient.Tests;
+namespace Arbeidstilsynet.Brevgenerator.Client.Tests;
 
 public class BrevgeneratorKlientTests
 {