+++ nuget-client/BrevgeneratorKlient/Model/BrevgeneratorSecret.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Arbeidstilsynet.Brevgenerator.Klient.Model;
+namespace Arbeidstilsynet.Brevgenerator.Client.Model;
 
 /// <summary>
 /// Representerer strukturen til OAuth2 client credentials secret.