+++ nuget-client/BrevgeneratorKlient/CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 4.0.0
 
 - Publiserer nå til `nuget.org` i stedet for Azure DevOps
-- Omdøpt pakke fra `AT.Brevgenerator.Klient` til `Arbeidstilsynet.Brevgenerator.Klient`
-- Omdøpt namespace fra `AT.Brevgenerator.Klient` / `AT.Brevgenerator.Klient.Model` til `Arbeidstilsynet.Brevgenerator.Klient` / `Arbeidstilsynet.Brevgenerator.Klient.Model`
+- Omdøpt pakke fra `AT.Brevgenerator.Klient` til `Arbeidstilsynet.Brevgenerator.Client`
+- Omdøpt namespace fra `AT.Brevgenerator.Klient` / `AT.Brevgenerator.Klient.Model` til `Arbeidstilsynet.Brevgenerator.Client` / `Arbeidstilsynet.Brevgenerator.Client.Model`
+- Omdøpt `BrevgeneratorKlient` til `BrevgeneratorClient`
 
 ## 3.3.0